Adds 'name' param to importlayers
Adds a new option to the `importlayers` command for `manage.py` which allows the layer's name to be specified directly instead of being derived from the title.  This allows for the `importlayers` command to be used to update a layer's title and data without creating a new layer each time the title and/or filename changes.